% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{apply.bias.corrections}
\alias{apply.bias.corrections}
\title{Apply bias corrections to model outputs}
\usage{
apply.bias.corrections(mod, bc)
}
\arguments{
\item{mod}{Data frame of model outputs}

\item{bc}{Vector of regional bias correction factors}
}
\value{
A data frame with bias corrected values
}
\description{
Apply bias corrections to model outputs
}
