% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{calc.elas.actual}
\alias{calc.elas.actual}
\title{Calculate actual elasticities using numerical derivatives.}
\usage{
calc.elas.actual(Ps, Pn, Y, params, basedata = NULL)
}
\arguments{
\item{Ps}{Staple food prices}

\item{Pn}{Nonstaple food prices}

\item{Y}{Per-capita income}

\item{params}{Model parameters.  See description in \code{\link{food.dmnd}}.}

\item{basedata}{Model results for the base values of Ps, Pn, and Y.}
}
\value{
data frame with elasticities
}
\description{
Given a set of prices and incomes, and model parameters,calculate the
elasticities using numerical derivatives.  Optionally, you can pass the model
results for the base values, if you've already calculated them.
}
\details{
The inputs Ps, Pn, and Y can be vectors, but if they are, they must all be
the same length
}
