% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{calc.hicks.actual}
\alias{calc.hicks.actual}
\title{Calculate the actual Hicks elasticities using the Slutsky equation.}
\usage{
calc.hicks.actual(eps, alpha.s, alpha.n, alpha.m)
}
\arguments{
\item{eps}{Elasticity values calculated by \code{\link{calc.elas.actual}}}

\item{alpha.s}{Budget fraction for staples}

\item{alpha.n}{Budget fraction for nonstaples}

\item{alpha.m}{Budget fraction for materials}
}
\value{
dataframe with hicks elasticities
}
\description{
Given actual price and income elasticities (i.e., calculated using finite
difference derivatives), compute the corresponding Hicks elasticities.
}
