% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc.pop.weight}
\alias{calc.pop.weight}
\title{Calculate a weight factor based on the population.}
\usage{
calc.pop.weight(input.data)
}
\arguments{
\item{input.data}{Data frame of food demand input data from FAO}
}
\value{
Data frame of input data with a population weight column added.
}
\description{
These weight factors can be used to give high population regions more
influence in the model fit relative to low population regions.
}
