% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{calc1eps}
\alias{calc1eps}
\title{Calculate the exponents in the demand equation.}
\usage{
calc1eps(alpha.s, alpha.n, eta.s, eta.n, xi)
}
\arguments{
\item{alpha.s}{Budget fraction for staple foods.}

\item{alpha.n}{Budget fraction for nonstaple foods.}

\item{eta.s}{Income elasticity for staple foods.}

\item{eta.n}{Income elasticity for nonstaple foods.}

\item{xi}{Matrix of Hicks elasticities}
}
\description{
This is an approximation to the Slutsky equations, inasmuch as we use these
as the exponents directly, instead of solving for exponents that produce
these values as the elasticities.
}
