% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{calc1q}
\alias{calc1q}
\title{Calculate demand quantities for a single set of inputs}
\usage{
calc1q(Ps, Pn, Y, eps, Ysterm, Ynterm, Acoef, psscl, pnscl)
}
\arguments{
\item{Ps}{Price of staple foods}

\item{Pn}{Price of nonstaple foods}

\item{Y}{Per-capita income}

\item{eps}{Matrix of Marshall elasticities}

\item{Ysterm}{Income term in the demand equation for staples}

\item{Ynterm}{Income term in the demand equation for nonstaples}

\item{Acoef}{Leading multiplier parameter.}
}
\description{
This is a helper function for the \code{mapply} calls in
\code{\link{food.dmnd}}
}
