% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create.xval.data}
\alias{create.xval.data}
\title{Create and save validation data sets.}
\usage{
create.xval.data(alldata, outdir = NULL)
}
\arguments{
\item{alldata}{Data frame of all food demand input data from FAO.}

\item{outdir}{Directory to write output into.  If omitted, don't write output
to files.}
}
\value{
List of lists of data frames.  The first list has two datasets:
\code{xval.byyear} and \code{xval.byrgn}.  Each of those datasets has a
\code{Testing} and a \code{Training} set.
}
\description{
We create two such sets:
\enumerate{
 \item{Split by year:  year > 2000 goes in the testing set.}
 \item{Split by region:  10 randomly-selected regions go in the testing set.}
}
}
\details{
Data are returned in a list.  If outdir is specified, they will
also be written to output files in that directory.

The original code for randomly (but repeatably) selecting regions for the
regional testing set was:
\code{
    set.seed(8675309)
    test.rgns <- sample(unique(as.character(alldata$GCAM_region_name)), 10)
}
However, changes in certain library code (e.g. dplyr) have occasionally
caused the order of region names to change in the output, which causes
the regions selected for the test set to change.  Worse, these changes
depend on the version of the library installed, meaning that two users
running the same version of this code could get different results.  To
prevent this, we now set the testing regions explicitly to the set that
we got from the procedure above at the time we submitted the first paper.
}
