% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ambrosia_parameters.R
\name{create_dataset_for_parameter_fit}
\alias{create_dataset_for_parameter_fit}
\title{Create dataset with observational error for log likelihood calculation using clustering.}
\usage{
create_dataset_for_parameter_fit(
  min_price_pd = 20,
  min_cal_fd = 1000,
  min_clusters = 300,
  lower_limit_sigma = 0.01,
  data = NULL
)
}
\arguments{
\item{min_cal_fd}{Minimum calories for the food demand model}

\item{min_clusters}{Minimum number of clusters to be generated by clustering algoritm. It is recommended to not lowewr this parameter below 20.}

\item{lower_limit_sigma}{Lower limit for sigma values calculated}

\item{data}{A data.frame or data.table with the raw data. Data should contain following names,

s_cal_pcap_day_thous (Containing 1000 calories per capita per day for staples)

ns_cal_pcap_day_thous (Containing 1000 calories per capita per day for non-staples)

gdp_pcap_thous (Containing GDP per capita in thous USD)

s_usd_p1000cal (Price of 1000 calories for staples per person per day)

ns_usd_p1000cal (Price of 1000 calories for non-staples per person per day)}

\item{min_price_fd}{Highest possible price of staples and non-staples}

\item{original_param_vector}{Original parameter vector to be used as the starting point for the optimization.These parameters are taken from Edmonds et al 2017}

\item{optim_method}{The optimization method to be used for maximization of the log likelihood. The default is set to BFGS}
}
\value{
A dataframe  called Processed_Data_for_MC with the following columns

sig2Qn- which is the observational error for non-staples

sig2Qs- which is the observational error for staples
}
\description{
Create dataset with observational error for log likelihood calculation using clustering.
}
\author{
KBN 2020
}
