% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{data_used_for_original_param}
\alias{data_used_for_original_param}
\title{Dataset used for original parameter estimation}
\format{
A data frame with data used for original params
}
\usage{
data(data_used_for_original_param)
}
\description{
Data used for original parameters estimated (output in mc_example)
}
\examples{
data(data_used_for_original_param)
}
\references{
Data used for original parameters estimated (output in mc_example)
}
\keyword{dataset}
