% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{food.dmnd.byincome}
\alias{food.dmnd.byincome}
\title{Tabulate food demand by per-capita-income}
\usage{
food.dmnd.byincome(obsdata, params, region = NULL)
}
\arguments{
\item{obsdata}{Data frame of observed prices and incomes.}

\item{params}{Model parameter structure.  See notes in
\code{\link{food.dmnd}}.}
}
\description{
Tabulate staple and nonstaple demand as a function of per-capita GDP.  This
function uses observed prices to
}
