% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{food.dmnd.byyear}
\alias{food.dmnd.byyear}
\title{Tabulate food demand by year for a model.}
\usage{
food.dmnd.byyear(obsdata, params, bc = NULL, region = NULL)
}
\arguments{
\item{obsdata}{Table of observed prices and incomes.}

\item{params}{Model parameter structure, as described in
\code{\link{food.dmnd}}}

\item{bc}{Vector of regional bias correction factors.  Optional. If
omitted, no bias correction will be applied.}

\item{region}{Region to apply the analysis to.}
}
\description{
Tabulate food demand by year for in input model using the observed prices and
incomes for a given region.  If \code{region} is \code{NULL}, do it for all
regions and concatenate the result.
}
\details{
This function is intended to generate model predictions that can be compared
to observed consumption.
}
