% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{food.dmnd}
\alias{food.dmnd}
\title{Calculate food demand using the Edmonds, et al. model.}
\usage{
food.dmnd(
  Ps,
  Pn,
  Y,
  params,
  rgn = NULL,
  staples_FE = 0,
  bias_adder_s = 0,
  bias_adder_ns = 0
)
}
\arguments{
\item{Ps}{Vector of staple food prices.}

\item{Pn}{Vector of nonstaple food prices.}

\item{Y}{Vector of per capita income.}

\item{params}{Model parameters structure (see details)}

\item{rgn}{Optional name for this calculation. If provided, the region will
be included as an extra column in the output data frame.}

\item{staples_FE}{Fixed effects term for staples. set to 0 by default}

\item{bias_adder_s}{Bias term for staples. Set to 0 by default}

\item{bias_adder_ns}{Bias term for non-staples. Set to 0 by default}
}
\description{
The Edmonds model divides food consumption into two categories,
\emph{staples}, which represent basic foodstuffs, and \emph{nonstaples},
which represent higher-quality foods.  Demand for staples increases at low
income, but eventually peaks and begins to decline with higher income.
Demand for nonstaples increases with income over all income ranges; however,
total (staple + nonstaple) demand saturates asymptotically at high income.
}
\section{Arguments}{


Ps and Pn are food prices (staple food price and nonstaple price) in
international dollars per 1000 (dietary) calories.  Y is per-capita GDP in
international dollars. Ps, Pn, Y may be vectors but must all be the same
length.

Params is a structure:
\describe{
\item{xi}{2x2 array of the xi elasticities}
\item{A}{Leading coefficients in the quantity calculations}
\item{yfunc}{Length-2 list of functions giving Y^eta(Y) (see note below)}
\item{Pm}{Price of ``materials''. Loosely speaking, this parameter controls
how valuable food is relative to everything else in the economy.}
}

Note that we don't need elasticity parameters for the materials component
because we calculate Qm as a residual.  That is, whatever portion of
household budgets is not spent on food is by definition spent on materials.
}

\section{Output}{


The return value is a data frame with the following elements.
\describe{
 \item{Qs}{Quantity for staple foods (S)}
 \item{Qn}{Quantity for nonstaple foods (N)}
 \item{Qm}{Quantity for ``materials'' (M).  This quantity
represents an aggregate of everything else besides food that consumers buy.}
 \item{alpha.s}{Budget fraction for S}
 \item{alpha.n}{Budget fraction for N}
 \item{alpha.m}{Budget fraction for M}
}
Demand for staple and nonstaple foods are given in thousands of dietary
calories per person per day.  Units for materials demand are unspecified.
}

\section{Income Elasticity Functions}{


For one of the functional forms used for the income behavior, eta(Y), eta
blows up, but Y^(eta(Y)) is well behaved.  Therefore, the eta functions need
to be able to calculate not just eta(Y), but Y^(eta(Y)), so they can handle
the limiting cases.
}

