% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-fcns.R
\docType{data}
\name{guiconst}
\alias{guiconst}
\alias{xidefault}
\alias{elasmin}
\alias{elasmax}
\alias{elasstep}
\alias{etastep}
\alias{spacer}
\title{Constants used in the GUI}
\format{
An object of class \code{numeric} of length 3.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{html} (inherits from \code{character}) of length 1.
}
\usage{
xidefault

elasmin

elasmax

elasstep

etastep

spacer
}
\description{
Constants used in the GUI

xidefault : Default xi values

elasmin : Minimum elasticity for the slider bar

elasmax : Maximum elasticity for the slider bar

elasstep : Elasticity bar step size

etastep : Eta slider bar step size

spacer : Spacer for labels
}
\keyword{datasets}
\keyword{internal}
