% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-fcns.R
\name{guiwidgets}
\alias{guiwidgets}
\alias{xi.matrix.input}
\alias{y0.input.box}
\alias{column.input.table}
\alias{eta.selector}
\title{Input widgets for food demand GUI app}
\usage{
xi.matrix.input()

y0.input.box()

column.input.table(inputids, defvals, min, max, step,
  labels = c("Staple", "Nonstaple"))

eta.selector(id, label2 = "\\\\(\\\\eta=f(Y)\\\\)", sel = 1)
}
\arguments{
\item{inputids}{Character vector of parameter identifiers}

\item{defvals}{Vector of default values}

\item{min}{Minimum value}

\item{max}{Maximum value}

\item{step}{Slider step size}

\item{labels}{Character vector of labels}
}
\description{
Input widgets for food demand GUI app
}
\section{Functions}{
\itemize{
\item \code{xi.matrix.input}: Input boxes for xi matrix

\item \code{y0.input.box}: Input boxes for y0

\item \code{column.input.table}: Draw input grid for other inputs

\item \code{eta.selector}: Draw selector widget for eta
}}

\keyword{internal}
