% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{lamks2nu1y0}
\alias{lamks2nu1y0}
\title{Convert the lambda and ks parameters to nu1 and y0}
\usage{
lamks2nu1y0(df)
}
\arguments{
\item{df}{Data frame of lambda and ks parameters.}
}
\description{
The staple income elasticity has two formulations of its parameters.  The one
that uses nu1 (income elasticity at Y==1) and y0 (value of Y for which the
elasticity is zero) is more intuitive, but it is incomplete.  That is, theere
are some valid models that simply cannot be expressed this way.  The model
can alternatively be expressed in terms of the lambda and ks parameters in
its equations.  This function allows an easy conversion from the latter
representation to the former.
}
