% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand-plots.R
\name{make.byincome.plot}
\alias{make.byincome.plot}
\title{Plot model results by per-capita income}
\usage{
make.byincome.plot(obsdata, params, region = NULL)
}
\arguments{
\item{obsdata}{Data frame of observed food demand data}

\item{params}{Model parameter structure}

\item{region}{Regions to include in the plot.}
}
\description{
Plot model results by per-capita income
}
