% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand-plots.R
\name{make.byyear.plot}
\alias{make.byyear.plot}
\title{Plot model results by year}
\usage{
make.byyear.plot(byyear.data, pltrgn = NULL)
}
\arguments{
\item{byyear.data}{Output from \code{\link{food.dmnd.byyear}}}

\item{pltrgn}{Region to plot}
}
\value{
Plot of model results by year
}
\description{
Plot model results by year
}
