% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand-plots.R
\name{make.demand.plot}
\alias{make.demand.plot}
\title{Plot staple, nonstaple, and total demand output from the model}
\usage{
make.demand.plot(alldata, xdata, xlabel, max.yval = NULL)
}
\arguments{
\item{alldata}{Data frame of output returned from \code{\link{food.dmnd}}.}

\item{xdata}{Vector of values for the x-axis.}

\item{xlabel}{Character string to use for the x-axis label}

\item{max.yval}{Maximum value to display on the y-axis.  See details.}
}
\description{
The plot will have quantity on the y-axis.  The x-axis data is passed in as a
separate vector, so it can be any of the input values used in the model (or
even in principle a variable that is only indirectly related to the input
values).
}
