% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand-mc.R
\name{mc.food.dmnd.byyear}
\alias{mc.food.dmnd.byyear}
\title{Compute food demand by year}
\usage{
mc.food.dmnd.byyear(obsdata, x, regions = NULL)
}
\arguments{
\item{obsdata}{Data frame of observed food demand}

\item{x}{Vector of model parameters in lambda-ks format.}

\item{regions}{Vector of regions to include in the computation.  If NULL,
include all.}
}
\value{
Data frame with results of food demand by year
}
\description{
The input parameters must be in the Monte Carlo formulation (i.e., lambda-ks,
\emph{not} nu1-y0 format
}
