% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{mc_data_example}
\alias{mc_data_example}
\title{Sample dataset of mc data}
\format{
A data frame with data created using an MCMC approach
}
\usage{
data(mc_data_example)
}
\description{
Data created using an MCMC approach
}
\keyword{datasets}
