% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpar-analysis.R
\name{mcparam.clip.tails}
\alias{mcparam.clip.tails}
\title{Filter a Monte Carlo distribution by quantiles}
\usage{
mcparam.clip.tails(mc.data, qlo = 0.01, qhi = 0.99)
}
\arguments{
\item{mc.data}{Data frame of Monte Carlo output}

\item{qlo}{Lower quantile for filtering}

\item{qhi}{Upper quantile for filtering.}
}
\value{
Logical vector with \code{TRUE} for rows that are in the main body of
the distribution, \code{FAlSE} for those that aren't.
}
\description{
This filters on all parameter simultaneously, so any sample that is outside
the requested quantile range in the marginal distribution of any parameter
will be excluded.
}
\details{
The log-posterior column is handled a bit differently; it isn't filtered on
the high end, just the low end.

The purpose of this function was to trim values far out on the tails that
made the plot scales impossible to read.
}
