% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpar-analysis.R
\name{mcparam.density}
\alias{mcparam.density}
\title{Create a density plot for all of the MC variables}
\usage{
mcparam.density(mc.data)
}
\arguments{
\item{mc.data}{Data frame of Monte Carlo output.}
}
\value{
A density plot
}
\description{
Create a density plot for all of the MC variables
}
