% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpar-analysis.R
\name{mcparam.itercount}
\alias{mcparam.itercount}
\title{Get the iteration count for a monte carlo dataset.}
\usage{
mcparam.itercount(niter, nproc, npset)
}
\arguments{
\item{niter}{Total number of iterations in the mcpar loop.}

\item{nproc}{Number of processors allocated to the calculation.}

\item{npset}{Number of parameter sets per processor}
}
\description{
This is only necessary for the C++ MC code, due to some limitations in the
way it records its output.
}
