% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpar-analysis.R
\name{mcparam.ML}
\alias{mcparam.ML}
\title{Get the maximum a-posteriori (MAP) parameters}
\usage{
mcparam.ML(mc.data)
}
\arguments{
\item{mc.data}{Data frame of Monte Carlo output}
}
\value{
A vector of parameters which yield the highest log likelihood.
}
\description{
The MAP parameters are those that produced the largest posterior probability
density.
}
