% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpar-analysis.R
\name{mcparam.sample}
\alias{mcparam.sample}
\title{Sample the MC results using bootstrap sampling.}
\usage{
mcparam.sample(mc.data, nsamp = 100, func = NULL)
}
\arguments{
\item{mc.data}{Data frame of Monte Carlo output}

\item{nsamp}{Number of samples to draw}

\item{func}{Optional function to apply to the samples drawn from the MC
distribution.}
}
\value{
Data frame or list (see details)
}
\description{
Optionally, apply a function to the sampled values.
}
\details{
If \code{func} is \code{NULL}, the return value will be a data frame of
sampled parameter values.  Otherwise the return value will be a list with the
data frame just described in the first element and the output of \code{func}
in the second.
}
