% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{merge.trn.tst}
\alias{merge.trn.tst}
\title{Create a merged dataset with training and test data, each labeled accordingly}
\usage{
\method{merge}{trn.tst}(obs.trn, obs.tst)
}
\arguments{
\item{obs.trn}{Data frame of observations in the training set.}

\item{obs.tst}{Data frame of observations in the testing set.}
}
\value{
merged dataframe
}
\description{
Create a merged dataset with training and test data, each labeled accordingly
}
