% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpar-analysis.R
\name{namemc}
\alias{namemc}
\title{Return the list of names for the parameters in the model.}
\usage{
namemc(nparam = 9)
}
\arguments{
\item{nparam}{Number of parameters.  Legal values are either 8 or 9.}
}
\value{
A vector of parameter names for the selected version of the model.
}
\description{
Supports both the 7 and 9 parameter version.  Adds the "LL" tag to the end to
cover the log likelihood column appened by the monte carlo code.
}
\details{
The 9 parameter version of the model is the canonical version (and the one
published in the paper)
}
