% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand-mc.R
\docType{data}
\name{plohi}
\alias{plohi}
\title{Recommended parameter limits for the model.}
\format{An object of class \code{matrix} with 2 rows and 9 columns.}
\usage{
plohi
}
\description{
Model parameters outside of these ranges are likely to produce models that
are numerically ill-behaved.  The first row of the matrix has recommended
minimum values; the second has recommended maximum values.
}
\details{
In other functions we talk about the "8-parameter" and "9-parameter" versions
of the model.  The 8-parameter version was used in early tests and is now
deprecated; therefore, these recommendations apply only to the 9-parameter
version.
}
\keyword{datasets}
