% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\name{prepare.obs}
\alias{prepare.obs}
\title{Prepare observations for use in the model.}
\usage{
prepare.obs(obs)
}
\arguments{
\item{obs}{Data frame of observed data.}
}
\description{
Convert units on prices and rename columns with long, wordy names to make
them easier to work with.
}
