% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{processed_data_example}
\alias{processed_data_example}
\title{Sample dataset of processed data}
\format{
A data frame with data processed using create_dataset_for_parameter_fit()
}
\usage{
data(processed_data_example)
}
\description{
Data processed using create_dataset_for_parameter_fit()
}
\keyword{datasets}
