% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpar-analysis.R
\name{read.mc.data}
\alias{read.mc.data}
\title{Read the Monte Carlo results file}
\usage{
read.mc.data(filename, varnames = namemc())
}
\arguments{
\item{filename}{Name of the file where the results are stored}

\item{varnames}{Names of the variables stored in the file (the MC output
format didn't have column headers). If omitted, a default set of names is
supplied.}
}
\value{
A table with parameter names in accordance with the food demand model
}
\description{
Read the Monte Carlo results file
}
