% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{runapp}
\alias{runapp}
\title{Launch the interactive GCAM food demand model}
\usage{
runapp()
}
\description{
The interactive model allows users to adjust model parameters and see how
they affect the model output.  To run the interactive model you must have the
R "shiny" package installed.
}
