% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand.R
\docType{data}
\name{testdata}
\alias{testdata}
\alias{y.vals}
\alias{Ps.vals}
\alias{Pn.vals}
\alias{Pm.vals}
\alias{samp.params}
\alias{x1}
\alias{x0}
\title{Sample values for the demand model.}
\format{An object of class \code{numeric} of length 20.}
\usage{
y.vals

Ps.vals

Pn.vals

Pm.vals

samp.params

x1

x0
}
\description{
These variables give examples of the data structures used in the model, as
well as baseline price, GDP and parameter values that should produce
reasonable results.

y.vals: Logarithmically-spaced pcGDP values

Ps.vals: Evenly spaced Ps values

Pn.vals: Evenly spaced Pn values

Pm.vals: Evenly spaced Pm values

samp.params: Example model parameter structure

x1: Monte carlo parameter vector.  This vector encodes the same model as
\code{samp.params}, but using the Monte Carlo version of the parameter
formulation for the staple income elasticity model.

x0: Parameters used to generate the test data
}
\keyword{datasets}
