% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{training_data}
\alias{training_data}
\title{Sample dataset of training data for ambrosia}
\format{
A data frame with sample data used within examples for ambrosia
}
\usage{
data(training_data)
}
\description{
Sample data used within examples for ambrosia
}
\examples{
data(training_data)
}
\references{
Created using the Process_Demand_Data.R script in the scripts folder
}
\keyword{dataset}
