% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-demand-mc.R
\name{vec2param}
\alias{vec2param}
\title{Convert a vector of parameters into a params structure.}
\usage{
vec2param(x)
}
\arguments{
\item{x}{Vector of model parameters, using the Monte Carlo formulation.}
}
\value{
Parameter structure suitable for use in \code{\link{food.dmnd}}.
}
\description{
We assume that if you're using this you are doing an Monte Carlo calculation,
so we set the parameters of eta.s accordingly.  We also look at the number of
parameters passed in.  If it is 8, we assume you want etas = constant.  If
it's 9, we assume you want etas = eta.s(lambda, k).  If it's anything else,
we throw an error.
}
\details{
The parameters in the vector are:
 [A_s, A_n, xi_ss, xi_cross, xi_nn, nu1_n, lambda_s, k_s, Pm ]
xi_cross is used for both xi_sn and xi_ns, forcing them to be equal.

If there are only 8 parameters, then the first 6 are as above,
and the next to last is eta_s.
}
