% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{write.fddata}
\alias{write.fddata}
\title{Write food demand data to csv files}
\usage{
write.fddata(dlist, stem = "food-dmnd-price", byname = FALSE)
}
\arguments{
\item{dlist}{A list of data frames.  Notionally these are subsets of a larger
dataset, as generated by split(...).}

\item{stem}{String giving the stem for the generated filenames.}

\item{byname}{Flag indicating whether the generated filenames should use the
names in \code{dlist}, or be sequentially generated.}
}
\description{
Write a list of data frames to files, with filenames generated either from
the names of the list, or by sequential numbering.
}
\details{
I'll be honest with you.  I really don't remember what I was using this
function for.  There is nothing specific to the food demand data in it, and
it doesn't appear to be used anywhere else.  If I don't rediscover its
purpose in the next round of updates, I'm going to delete it.
}
